Move CodeGen_Hexagon to internal linkage.
